Condensing YAQL unit tests into a single file for now
