--- st2mistral/tests/unit/test_yaql_time.py
@@ -1,31 +1,0 @@
-# Licensed to the StackStorm, Inc ('StackStorm') under one or more
-# contributor license agreements.  See the NOTICE file distributed with
-# this work for additional information regarding copyright ownership.
-# The ASF licenses this file to You under the Apache License, Version 2.0
-# (the "License"); you may not use this file except in compliance with
-# the License.  You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from st2mistral.tests.unit import test_filter_base as base
-
-from yaql.language import factory
-YAQL_ENGINE = factory.YaqlFactory().create()
-
-
-class TestCaseYaqlTime(base.YaqlFilterTestCase):
-
-    def test_to_human_time_from_seconds(self):
-
-        result = YAQL_ENGINE('to_human_time_from_seconds(12345)').evaluate(
-            context=self.get_yaql_context({})
-        )
-        self.assertEqual(result, '3h25m45s')
-
-
